+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Next Version
 ### Misc
 - Grouped benchmarks in a pattern that should make more sense.
+- Added pydantic docs to `README.md`
+- Added graphs and benchmark results.
 ## 0.2.1 - 12/6/2021
 ### Fixed
 - Depth limit is now enforced for `ormsgpack.unpackb` - function should be safe for use now.